requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cleanup==6.0.0
django-crispy-forms==1.14.0
persian-tools==0.0.10
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
